+++ lib/core/zenith/zenith_core.dart
@@ -1,10 +1,14 @@
 export 'async_value.dart';
+export 'auth_state.dart';
 export 'computed_node.dart';
+export 'persisted_node.dart';
 export 'zenith_concurrency.dart';
 export 'zenith_container.dart';
 export 'zenith_key.dart';
+export 'zenith_middleware.dart';
 export 'zenith_node.dart';
 export 'zenith_observer.dart';
 export 'zenith_selector.dart';
 export 'zenith_stream.dart';
+export 'zenith_storage.dart';
 export 'zenith_widgets.dart';